refactor: changed config name for specificity and removed unused imports
