+++ swagger_generator/generator_config.py
